+++ packages/@expo/metro-runtime/src/error-overlay/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+// TODO: This will break tree shaking due to how we transpile this package.
 import { Platform } from 'react-native';
 
 import ErrorToastContainer from './toast/ErrorToastContainer';
@@ -11,10 +12,12 @@ if (!global.setImmediate) {
   };
 }
 
-if (process.env.NODE_ENV === 'development' && Platform.OS === 'web') {
-  // Stack traces are big with React Navigation
+if (process.env.NODE_ENV === 'development') {
+  if (Platform.OS === 'web') {
+    // Stack traces are big with React Navigation
 
-  require('./LogBox').default.install();
+    require('./LogBox').default.install();
+  }
 }
 
 export function withErrorOverlay(Comp: React.ComponentType<any>) {