+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/views/ViewManagerWrapperDelegate.kt
@@ -73,7 +73,7 @@ class ViewManagerWrapperDelegate(internal var moduleHolder: ModuleHolder) {
       val key = iterator.nextKey()
       expoProps[key]?.let { expoProp ->
         try {
-          expoProp.set(propsMap.getDynamic(key), view)
+          expoProp.set(propsMap.getDynamic(key), view, moduleHolder.module._appContext)
         } catch (exception: Throwable) {
           // The view wasn't constructed correctly, so errors are expected.
           // We can ignore them.