+++ packages/expo-font/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### 💡 Others
 
+- Updated types for server functions. ([#23911](https://github.com/expo/expo/pull/23911) by [@EvanBacon](https://github.com/EvanBacon))
 - Migrated `FontLoaderModule` to use Expo Modules API. ([#24015](https://github.com/expo/expo/pull/24015) by [@lukmccall](https://github.com/lukmccall))
 
 ## 11.5.1 — 2023-08-02